+++ library-events-consumer-v2/src/main/java/com/learnkafka/repository/LibraryEventRepository.java
@@ -3,6 +3,6 @@ package com.learnkafka.repository;
 import com.learnkafka.domain.LibraryEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LibraryEventRepository extends JpaRepository<LibraryEvent, Integer> {
+public interface LibraryEventRepository extends JpaRepository<LibraryEvent, Long> {
 }
